add tests & cs workflows
