+++ .github/workflows/cs-tests.yml
@@ -1,0 +1,63 @@
+name: Static Code Anylzer & Tests
+on:
+  push:
+
+jobs:
+    phpcpd:
+        name: phpcpd
+        runs-on: ubuntu-latest
+        steps:
+            - name: Setup PHP
+              uses: shivammathur/setup-php@2.12.0
+              with:
+                  php-version: '7.4'
+                  tools: composer:v2, phpcpd
+            - name: Checkout
+              uses: actions/checkout@v2.3.4
+            - run: composer install --prefer-dist
+            - run: phpcpd ./ --exclude=vendor --exclude=bin --exclude=tests
+
+    php-cs-fixer:
+        name: php-cs-fixer
+        runs-on: ubuntu-latest
+        steps:
+            - name: Setup PHP
+              uses: shivammathur/setup-php@2.12.0
+              with:
+                  php-version: '7.4'
+                  tools: composer:v2, php-cs-fixer, cs2pr
+            - name: Checkout
+              uses: actions/checkout@v2.3.4
+            - run: composer install --prefer-dist
+            - run: php-cs-fixer fix --config=.php-cs-fixer.dist.php -v --dry-run --using-cache=no --format=checkstyle | cs2pr
+
+    psalm:
+        name: psalm
+        runs-on: ubuntu-latest
+        steps:
+            - name: Setup PHP
+              uses: shivammathur/setup-php@2.12.0
+              with:
+                  php-version: '7.4'
+                  tools: composer:v2, psalm
+            - uses: actions/checkout@v2
+            - run: composer install --prefer-dist
+            - run: psalm --config=psalm.xml --no-cache --output-format=github
+
+    tests:
+        runs-on: ubuntu-latest
+        steps:
+            - name: Setup PHP
+              uses: shivammathur/setup-php@2.12.0
+              with:
+                  php-version: '7.4'
+                  tools: composer:v2
+            - uses: actions/checkout@v2
+            - uses: php-actions/composer@v6
+              with:
+                php_version: 7.4
+            - run: ./vendor/bin/phpunit
+
+#              run: XDEBUG_MODE=coverage ./vendor/bin/phpunit
+#            - name: Update Coveralls
+#              run: php vendor/bin/php-coveralls -v
